REFACTORED: Pop up animation extracted as a custom hook, applied to all pop up components
