+++ components/pop-up/accept-alert.tsx
@@ -1,10 +1,9 @@
 import { Button } from "../ui/button";
 import { Animated, Text } from "react-native";
-import { useEffect, useRef } from "react";
-import { useState } from "react";
 import { Card, CardContent } from "../ui/card";
 import { Portal } from "@rn-primitives/portal";
 import { StyleSheet } from "react-native";
+import { useAnimatedModal } from "~/hooks/use-animated-modal";
 
 interface AcceptAlertProps {
   visible: boolean;
@@ -12,50 +11,11 @@ interface AcceptAlertProps {
 }
 
 export default function AcceptAlert({ visible, setOpen }: AcceptAlertProps) {
-  const [mounted, setMounted] = useState(visible);
-  const backdropOpacity = useRef(new Animated.Value(0)).current;
-  const cardOpacity = useRef(new Animated.Value(0)).current;
+  const { mounted, backdropOpacity, cardOpacity } = useAnimatedModal(visible);
 
   const handleClose = () => {
     setOpen(false);
   };
-
-  useEffect(() => {
-    let cancelled = false;
-    if (visible) {
-      setMounted(true);
-      Animated.parallel([
-        Animated.timing(backdropOpacity, {
-          toValue: 1,
-          duration: 250,
-          useNativeDriver: true,
-        }),
-        Animated.timing(cardOpacity, {
-          toValue: 1,
-          duration: 250,
-          useNativeDriver: true,
-        }),
-      ]).start();
-    } else if (mounted) {
-      Animated.parallel([
-        Animated.timing(cardOpacity, {
-          toValue: 0,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-        Animated.timing(backdropOpacity, {
-          toValue: 0,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-      ]).start(({ finished }) => {
-        if (!cancelled && finished) setMounted(false);
-      });
-    }
-    return () => {
-      cancelled = true;
-    };
-  }, [visible]);
 
   if (!mounted) return null;
