+++ components/pop-up/incorrect-password.tsx
@@ -1,8 +1,8 @@
-import React, { useEffect, useRef, useState } from "react";
 import { Animated, StyleSheet, Text } from "react-native";
 import { Portal } from "@rn-primitives/portal";
 import { Card, CardContent } from "~/components/ui/card";
 import { Button } from "~/components/ui/button";
+import { useAnimatedModal } from "~/hooks/use-animated-modal";
 
 interface IncorrectPasswordProps {
   visible: boolean;
@@ -13,50 +13,11 @@ export function IncorrectPassword({
   visible,
   setOpen,
 }: IncorrectPasswordProps) {
-  const [mounted, setMounted] = useState(visible);
-  const backdropOpacity = useRef(new Animated.Value(0)).current;
-  const cardOpacity = useRef(new Animated.Value(0)).current;
+  const { mounted, backdropOpacity, cardOpacity } = useAnimatedModal(visible);
 
   const handleClose = () => {
     setOpen(false);
   };
-
-  useEffect(() => {
-    let cancelled = false;
-    if (visible) {
-      setMounted(true);
-      Animated.parallel([
-        Animated.timing(backdropOpacity, {
-          toValue: 1,
-          duration: 250,
-          useNativeDriver: true,
-        }),
-        Animated.timing(cardOpacity, {
-          toValue: 1,
-          duration: 250,
-          useNativeDriver: true,
-        }),
-      ]).start();
-    } else if (mounted) {
-      Animated.parallel([
-        Animated.timing(cardOpacity, {
-          toValue: 0,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-        Animated.timing(backdropOpacity, {
-          toValue: 0,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-      ]).start(({ finished }) => {
-        if (!cancelled && finished) setMounted(false);
-      });
-    }
-    return () => {
-      cancelled = true;
-    };
-  }, [visible]);
 
   if (!mounted) return null;
