+++ components/pop-up/move-to-trash-alert.tsx
@@ -1,9 +1,9 @@
-import React, { useEffect, useRef, useState } from "react";
 import { Animated, StyleSheet, Text, View } from "react-native";
 import { Portal } from "@rn-primitives/portal";
 import { Card, CardContent } from "~/components/ui/card";
 import { Button } from "~/components/ui/button";
 import MaterialIcons from "@expo/vector-icons/MaterialIcons";
+import { useAnimatedModal } from "~/hooks/use-animated-modal";
 
 interface MoveToTrashAlertProps {
   visible: boolean;
@@ -16,46 +16,7 @@ export function MoveToTrashAlert({
   onConfirm,
   onCancel,
 }: MoveToTrashAlertProps) {
-  const [mounted, setMounted] = useState(visible);
-  const backdropOpacity = useRef(new Animated.Value(0)).current;
-  const cardOpacity = useRef(new Animated.Value(0)).current;
-
-  useEffect(() => {
-    let cancelled = false;
-    if (visible) {
-      setMounted(true);
-      Animated.parallel([
-        Animated.timing(backdropOpacity, {
-          toValue: 1,
-          duration: 250,
-          useNativeDriver: true,
-        }),
-        Animated.timing(cardOpacity, {
-          toValue: 1,
-          duration: 250,
-          useNativeDriver: true,
-        }),
-      ]).start();
-    } else if (mounted) {
-      Animated.parallel([
-        Animated.timing(cardOpacity, {
-          toValue: 0,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-        Animated.timing(backdropOpacity, {
-          toValue: 0,
-          duration: 200,
-          useNativeDriver: true,
-        }),
-      ]).start(({ finished }) => {
-        if (!cancelled && finished) setMounted(false);
-      });
-    }
-    return () => {
-      cancelled = true;
-    };
-  }, [visible]);
+  const { mounted, backdropOpacity, cardOpacity } = useAnimatedModal(visible);
 
   if (!mounted) return null;
